fix: clearning up #1
