+++ portal/two_factor/form.py
@@ -3,95 +3,18 @@ from time import time
 
 from django import forms
 from django.conf import settings
-from django.forms import Form, ModelForm
+from django.forms import Form
 from django.utils.translation import gettext_lazy as _
 from django_otp.forms import OTPAuthenticationFormMixin
 from django_otp.oath import totp
 from django_otp.plugins.otp_totp.models import TOTPDevice
 
-from two_factor.models import (
-    PhoneDevice,
-    get_available_methods,
-    get_available_phone_methods,
-)
 from two_factor.utils import totp_digits
-from two_factor.validators import validate_international_phonenumber
 
 try:
     from otp_yubikey.models import RemoteYubikeyDevice, YubikeyDevice
 except ImportError:
     RemoteYubikeyDevice = YubikeyDevice = None
-
-
-class MethodForm(forms.Form):
-    method = forms.ChoiceField(
-        label=_("Method"), initial="generator", widget=forms.RadioSelect
-    )
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.fields["method"].choices = get_available_methods()
-
-
-class PhoneNumberMethodForm(ModelForm):
-    number = forms.CharField(
-        label=_("Phone Number"), validators=[validate_international_phonenumber]
-    )
-    method = forms.ChoiceField(widget=forms.RadioSelect, label=_("Method"))
-
-    class Meta:
-        model = PhoneDevice
-        fields = (
-            "number",
-            "method",
-        )
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self.fields["method"].choices = get_available_phone_methods()
-
-
-class PhoneNumberForm(ModelForm):
-    # Cannot use PhoneNumberField, as it produces a PhoneNumber object, which cannot be serialized.
-    number = forms.CharField(
-        label=_("Phone Number"), validators=[validate_international_phonenumber]
-    )
-
-    class Meta:
-        model = PhoneDevice
-        fields = ("number",)
-
-
-class DeviceValidationForm(forms.Form):
-    token = forms.IntegerField(
-        label=_("Token"), min_value=1, max_value=int("9" * totp_digits())
-    )
-
-    error_messages = {
-        "invalid_token": _("Entered token is not valid."),
-    }
-
-    def __init__(self, device, **args):
-        super().__init__(**args)
-        self.device = device
-
-    def clean_token(self):
-        token = self.cleaned_data["token"]
-        if not self.device.verify_token(token):
-            raise forms.ValidationError(self.error_messages["invalid_token"])
-        return token
-
-
-class YubiKeyDeviceForm(DeviceValidationForm):
-    token = forms.CharField(label=_("YubiKey"), widget=forms.PasswordInput())
-
-    error_messages = {
-        "invalid_token": _("The YubiKey could not be verified."),
-    }
-
-    def clean_token(self):
-        self.device.public_id = self.cleaned_data["token"][:-32]
-        return super().clean_token()
 
 
 class TOTPDeviceForm(forms.Form):
